feat: implement OwnerLookupDropdown component and refactor AddContactModal
