+++ frontend/src/community/crm/constants/contactConstants.ts
@@ -1,5 +1,3 @@
-export const CONTACT_SEARCH_DEBOUNCE_DELAY = 500;
 export const DEFAULT_PAGE_SIZE = 10;
 export const DEFAULT_COMPANY_PAGE_SIZE = 100;
 export const ALL_COMPANIES = undefined;
-export const DEFAULT_LOOKUP_PAGE_SIZE = 50;
